server/requirements.txt
fastapi~=0.88.0
uvicorn~=0.20.0
python-decouple~=3.6
elasticsearch~=8.5.2
click~=8.1.3
